pyproject.toml
[project]
name = "llm-clean"
version = "0.1.0"
description = "Ontologically Sound Taxonomy Generation with Agentic Critique"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rdflib",
    "requests",
    "python-dotenv",
    "PyPDF2",
    "tqdm",
    "pandas",
    "tabulate",
    "numpy",
    "langchain-openai",
    "langchain-core",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
